Moved all arena-rewrite codes to default paths and merge into sw core.
Moved some stuffs into features folder.
