+++ src/larryTheCoder/arena/tasks/ArenaGameTick.php
@@ -26,25 +26,15 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace larryTheCoder\arenaRewrite\tasks;
+namespace larryTheCoder\arena\tasks;
 
-use larryTheCoder\arenaRewrite\Arena;
-use larryTheCoder\SkyWarsPE;
-use larryTheCoder\utils\Utils;
-use pocketmine\block\WallSign;
-use pocketmine\math\Vector3;
+
+use larryTheCoder\arena\Arena;
 use pocketmine\scheduler\Task;
-use pocketmine\Server;
-use pocketmine\tile\Sign;
 
-class SignTickTask extends Task {
-
-	/** @var Arena */
-	private $arena;
-	private $updateTime;
+class ArenaGameTick extends Task {
 
 	public function __construct(Arena $arena){
-		$this->arena = $arena;
 	}
 
 	/**
@@ -55,25 +45,6 @@ class SignTickTask extends Task {
 	 * @return void
 	 */
 	public function onRun(int $currentTick){
-		$this->updateTime++;
-		if($this->updateTime >= $this->arena->statusLineUpdate){
-			$vars = ['%alive', '%status', '%max', '&', '%world', '%prefix', '%name'];
-			$replace = [count($this->arena->getPlayers()), $this->arena->getStatus(), $this->arena->maximumPlayers, "§", $this->arena->arenaWorld, SkyWarsPE::getInstance()->getPrefix(), $this->arena->arenaName];
-			$level = Server::getInstance()->getLevelByName($this->arena->joinSignWorld);
-			if($level === null){
-				goto skipUpdate;
-			}
-			$tile = $level->getTile(new Vector3($this->arena->joinSignX, $this->arena->joinSignY, $this->arena->joinSignZ));
-			if($tile instanceof Sign){
-				$block = $tile->getLevel()->getBlock($tile);
-				if($block instanceof WallSign){
-					$vec = $block->getSide($block->getDamage() ^ 0x01);
-					$tile->getLevel()->setBlock($vec, Utils::getBlockStatus($this->arena));
-				}
-				$tile->setText(str_replace($vars, $replace, $this->arena->statusLine1), str_replace($vars, $replace, $this->arena->statusLine2), str_replace($vars, $replace, $this->arena->statusLine3), str_replace($vars, $replace, $this->arena->statusLine4));
-			}
-			skipUpdate:
-			$this->updateTime = 0;
-		}
+		// TODO: Implement onRun() method.
 	}
 }
