+++ src/larryTheCoder/arena/tasks/SignTickTask.php
@@ -26,15 +26,25 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace larryTheCoder\arenaRewrite\tasks;
+namespace larryTheCoder\arena\tasks;
 
-
-use larryTheCoder\arenaRewrite\Arena;
+use larryTheCoder\arena\Arena;
+use larryTheCoder\SkyWarsPE;
+use larryTheCoder\utils\Utils;
+use pocketmine\block\WallSign;
+use pocketmine\math\Vector3;
 use pocketmine\scheduler\Task;
+use pocketmine\Server;
+use pocketmine\tile\Sign;
 
-class ArenaGameTick extends Task {
+class SignTickTask extends Task {
+
+	/** @var Arena */
+	private $arena;
+	private $updateTime;
 
 	public function __construct(Arena $arena){
+		$this->arena = $arena;
 	}
 
 	/**
@@ -45,6 +55,25 @@ class ArenaGameTick extends Task {
 	 * @return void
 	 */
 	public function onRun(int $currentTick){
-		// TODO: Implement onRun() method.
+		$this->updateTime++;
+		if($this->updateTime >= $this->arena->statusLineUpdate){
+			$vars = ['%alive', '%status', '%max', '&', '%world', '%prefix', '%name'];
+			$replace = [count($this->arena->getPlayers()), $this->arena->getStatus(), $this->arena->maximumPlayers, "§", $this->arena->arenaWorld, SkyWarsPE::getInstance()->getPrefix(), $this->arena->arenaName];
+			$level = Server::getInstance()->getLevelByName($this->arena->joinSignWorld);
+			if($level === null){
+				goto skipUpdate;
+			}
+			$tile = $level->getTile(new Vector3($this->arena->joinSignX, $this->arena->joinSignY, $this->arena->joinSignZ));
+			if($tile instanceof Sign){
+				$block = $tile->getLevel()->getBlock($tile);
+				if($block instanceof WallSign){
+					$vec = $block->getSide($block->getDamage() ^ 0x01);
+					$tile->getLevel()->setBlock($vec, Utils::getBlockStatus($this->arena));
+				}
+				$tile->setText(str_replace($vars, $replace, $this->arena->statusLine1), str_replace($vars, $replace, $this->arena->statusLine2), str_replace($vars, $replace, $this->arena->statusLine3), str_replace($vars, $replace, $this->arena->statusLine4));
+			}
+			skipUpdate:
+			$this->updateTime = 0;
+		}
 	}
 }
