+++ src/larryTheCoder/features/kits/Kits.php
@@ -26,7 +26,7 @@
  * USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-namespace larryTheCoder\libs\kits;
+namespace larryTheCoder\features\kits;
 
 use larryTheCoder\SkyWarsPE;
 use onebone\economyapi\EconomyAPI;
